ex2

+++ modulo2/semana3/exsoma/Program.cs
@@ -4,10 +4,10 @@
     {
 
         Console.WriteLine("Digite o primeiro número: ");
-        int num1 = Convert.ToInt32(Console.ReadLine());
+        double num1 = Convert.ToInt32(Console.ReadLine());
 
         Console.WriteLine("Digite o segundo número: ");
-        int num2 = Convert.ToInt32(Console.ReadLine());
+        double num2 = Convert.ToInt32(Console.ReadLine());
 
         Console.WriteLine("Digite o terceiro número: ");
         double num3 = Convert.ToDouble(Console.ReadLine());
